requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
autopep8==2.0.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.1
fastapi==0.96.1
ffmpy==0.3.0
filelock==3.12.1
filetype==1.2.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.6.0
gradio==3.34.0
gradio_client==0.2.6
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.31.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.2
linkify-it-py==2.0.2
llvmlite==0.40.1rc1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
onnxruntime==1.15.0
opencv-python-headless==4.7.0.72
orjson==3.9.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.3
pooch==1.7.0
protobuf==4.23.2
pycodestyle==2.10.0
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyMatting==1.1.8
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rembg==2.0.43
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tifffile==2023.4.12
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.3
uvicorn==0.22.0
watchdog==3.0.0
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
